requirements.txt
streamlit
python-dotenv
langchain
langchain-huggingface
langchain-openai
langchain-pinecone
langchain-core
pinecone-client
openai
sentence-transformers
transformers
torch
numpy
regex
huggingface_hub